Handle sparse files in tarfiles
These are a bit weird, because the on-disk data does not match the
content in the tarball, and also the golang tar library auto-expands
it.  We could un-expand it and do the correct conversion, but thats a
lot of work for a rare corner case, so instead we just avoid delta:ing these.

Signed-off-by: Alexander Larsson <alexl@redhat.com>