Tweak how we use sizes to pick diff source
We now don't limit ourselves to sizes within 30% (because in testing
we had cases where it changed more and yet bsdiff helped). Instead
we compare the hits by size and pick the one closes in size.

However, we do *some* size changes, so if two files are wildly
dissimilar in size they are not condidered for diff:ing. Wildly
dissimilar for now means a factor of 10, and only for larger (> 64k)
files.

Signed-off-by: Alexander Larsson <alexl@redhat.com>